+++ src/types/index.ts
@@ -4,3 +4,4 @@ export * from "./nearest";
 export * from "./matrix";
 export * from "./isochrones";
 export * from "./provider";
+export * from "./authentication";